Respect sensor orientation with ignore-orientation-request
So the behavior follows the API definition of sensor orientation.
And the target to ignore is mainly for fixed-orientation.

Also optimize the cases of unset/unspecified/user that are also not
fixed-orientation. It can save hundreds of activity traversals by
shouldRespectOrientationRequestDueToPerAppOverride when launching
a regular app.

Fix: 432587525
Flag: EXEMPT bugfix
Test: DisplayContentTests#testOnDescendantOrientationRequestChanged
Change-Id: I890d539ebd8fc10d591066814f1a1de8b1f28b3d