Default summaries on for droid/nextfooders.
This change resets the XML attribute name, thus resetting the overall enabled setting on both bundles and summaries. As this would render the "is setting set by user" tracking for bundles outdated as well, it also resets that parameter. This will set both bundles and summaries to default ON, with bundles off for the work profile.

Bug: 433554352
Test: NotificationManagerServiceTest, NotificationAssistantsTest; manual by setting all settings opposite the default (bundles on for work profile but off overall, summaries off) and confirming that patching this change resets these settings.
Flag: android.app.nm_summarization_onboarding_ui
Change-Id: Idcd1b56ae63045732f9334a86af09606f4b6fad1