Make sure stableBounds during drag up-to-date
During a drag-resize operation, the stableBounds were not being
recalculated when the task was moved to a different display. This could
lead to incorrect resizing constraints.

This change ensures that stableBounds are updated when the task's
displayId at the end of a drag does not match the displayId from the
start, guaranteeing correct resize behavior with multiple-display
enabled.

Flag: com.android.window.flags.enable_drag_end_stable_bounds_reset
Bug: 421180735
Test: manual; atest
Change-Id: I714235c42bfaa34b8d73b0afd772ed7e6d677e41