[Bubbles] Run the hide-TaskView WCT after the current method execution
The WCT to hide TaskView may take significant time. If executed in the
removeView method call, it might result in missed frames. Before
ag/34380975, it was executed after the removeView call because
`startNextTransition` picks up a different wct in the queue (the CHANGE
transition caused by TaskView bounds update). After ag/34380975, because
of the elimination of the extra CHANGE transition, the hide-TaskView WCT
is executed synchronously on the animation callback, causing missed
frames. This CL posts the wct execution to the ShellHandler so that it
doesn't block the thread.

Bug: 431100601
Bug: 424205239
Test: atest TaskViewTransitionsTest BubbleControllerTest
Flag: EXEMPT bugfix
Change-Id: I8e25a3e8ce775e5564d2d49f3a6f844b6fb1866d