Remove task from desktop repo when moved to unsupported display
When a task active in the desktop is moved to a display that does not
support desktop mode, it becomes a non-freeform task. This change
ensures that we remove it from the `DesktopRepository` to prevent stale
state.

Bug: 424748132
Test: DesktopTaskChangeListenerTest
Flag: com.android.window.flags.move_to_next_display_shortcut_with_projected_mode
Change-Id: I67933c0367357e00de1e559568222e657f767e88