Only move to pinned task if activity non finishing
When HIERARCHY_OP_TYPE_MOVE_PIP_ACTIVITY_TO_PINNED_TASK
is being applied upon request from Shell, we should make sure
the PiP activity is still non finishing one.

This is possible because while Shell is handling transition
request-start, Core is free to accept and process calls from the app,
such as the finishing of the activity that was requested to PiP earlier.

Bug: 433490150
FLAG: EXEMPT bugfix
Test: manually force-stop the activity while swiping-to-pip in ges-nav
Change-Id: I685142d14543780228ce05b65c9162f2cbd378f5