Update non-inherited task surface crops during syncs
We generally want to let shell manage the surfaces of
tasks that it organizes. However, one of the transition
contracts is that the startTransaction, when applied,
should produce the "end state". This means that we want
WM to include surface changes during sync -- otherwise
if a change wasn't a direct participant of a transition,
but still changed, that change won't be represented.

Bug: 410684072
Test: resize splits with a aspect/size restricted task
Flag: com.android.window.flags.update_task_crop_in_sync
Change-Id: Ie2aad45380c9f7a5bdf4422b3be923f933e838fe