Update ProfcollectForwardingService broadcast as current user.
The broadcast from ProfcollectForwardingService to trigger a profile
upload was sent only to the system user. The ProfcollectUploadReceiver,
however, is not available for the system user, causing uploads to fail.

This change sends the broadcast as current user, ensuring the
ProfcollectUploadReceiver can process the upload request. This was
tested on an Android desktop device.

Bug: 434133383
Test: Ran `adb shell cmd jobscheduler run -f android 260817` and
observed 'ProfcollectUploadReceiver: Received upload intent' in logcat.
Flag: EXEMPT bugfix
Change-Id: Ibcc9cc5d7e19c1f954802eca1aefbc72449b6db4