Avoid scheduling duplicated events to check activity drawn
Bad apps could start activities repeatedly in a short period of time
and resulted in lots of transitions. Lots of events were then
scheduled to check the previous transitions whenever a new
transition is added. That generated too many messages and caused
the handler to be busy for several seconds.

Also avoid resuming the rest activities in the Task if the Task
is currently being removed.

Task#removeImmediately -> … -> ActivityRecord#removeImmediately
-> ActivityRecord#destroyed -> RWC#resumeFocusedTasksTopActivities

Bug: 428943578
Test: wm presubmit
Flag: EXEMPT bugfix
Change-Id: I9629ef1f802460ada55ddac5b1d2d91445a0fcb3