Posting show/hide call if view is running on a different thread
When we received a show/hide call via binder call from system_server, it was always directly sent to InsetsController. However, this did not take the possibility into account that the view is running on different thread.
This CL adds a check for that and, if needed, posts the show/hide call to the correct thread.

Fix: 407546107
Test: atest CtsInputMethodTestCases
Flag: EXEMPT bugfix
Change-Id: I99e89aa824f2063336784255f378b2a44b5b016e