Revert "[Floaty] Don't emit cancel event when effect is disabled"
This reverts commit 67c1a6b2817240d5a602379e586746641a3a4237.

Reason for revert: b/433842249

In some cases we actually want to emit a CANCEL_SQUEEZE event when the
squeeze effect is disabled. Namely, when the effect gets disabled while
an animation is in progress. This happens on tablets sometimes because
the active assistant changes frequently on tablets. While locked,
tablets have Google Assistant active and while unlocked tablets have
Gemini active. So whenever the device is locked or unlocked, the active
assistant changes.

Bug: 433842249
Test: SqueezeEffectInteractorTest
Test: Manual, i.e. verified on tablets that invocation effect never
      plays when turning on the display with the power button.
Change-Id: Ide6ec19dc74672ca0860aba3f74f4f285c53076b