Allow overriding default visibility for sensitive inputs
Allows overriding the default values of the settings that control the
visibility of sensitive inputs such as passwords, PINs and patterns via
config.xml.

This implementation is a temporary measure pending a permanent solution,
which is being tracked under b/417951523. The previously considered
approach, which would have determined the default based on the
availability of physical input device, was set aside to prioritize the
development of the permanent solution.

Bug: 339270220
Test: Manual
Test: atest LockPatternUtilsTest
Flag: com.android.internal.widget.flags.use_default_visibility_for_sensitive_inputs
Change-Id: I2a2a4261fe161a14c5093332c24dd6df5ecd9c29