Fix(Connected Displays): Show permission dialog on the correct display
The media projection permission dialog was always shown on the primary display.

This was because MediaProjectionPermissionActivity used getApplicationContext(), which is tied to the default display. The fix is to use createDisplayContext(getDisplay()) to ensure the dialog appears on the same display as the requesting app.

Test: Tested share screen dialog using meet app on connected display and
in split screen
Bug: 424091508
Flag: EXEMPT minor fix

Change-Id: I19d5ffda80d4677b306e3f2ebcb5e1a991bf1266