Only check the systemEntity of an admin if it is known to be a SystemAuthority.
Non-SystemAuthority admins will not have a system entity and should not be targeted by removeLocalPoliciesForSystemEntities. It could also cause an NPE to be thrown when `List#contains()` is called with a null systemEntity as the argument, depending on the list type that is passed in.

Test: atest DevicePolicyEngineTest
Bug: 434849370
Flag: android.app.supervision.flags.supervision_manager_apis

Change-Id: If674a7b852938484cd03596f42489740f696342e