Update VersionCollapser to consider minorVersion
The tests have been updated to use `AddString` instead of `AddSimple`, so that we can ensure that the correct configs were removed and renamed. They have also been refactored to be more readable.

This change also fixes a bug in the `ResTable_config` compare functions, where `version` was being compared. Since it is in a union with a struct of `sdkVersion` and `minorVersion`, the endianness determines which of the members is more significant. We now explicitly compare both fields to ensure consistency.

Bug: 383177182
Flag: EXEMPT Aconfig not supported on host tools
Test: atest aapt2_tests
Change-Id: Ib61918765da8864d26535c90d41df5a5d055d66b