[ADI][54/N] bring back AppNotInstalled dialog
Based on the latest UX design, we still need the AppNotInstalled dialog
to inform the user that the app was not installed after the user
declines in the ADI verification dialog. We also want to disable
clicking outside of the dialog so that the only way to dismiss the
dialog is to click on the button(s).

BUG: 360129657
FLAG: android.content.pm.verification_service
Test: manual

Change-Id: I98a3b849d224c2bba7b8049d65a0de280843f710