Correct 8196-bit RSA key typo for CertUtils and SigXml tests
The CertUtilsTest and SigXmlTest both use a root CA to sign
intermediate CAs which in turn sign leaf certificates that are
used to verify the functionality within CertUtils and SigXml.
This root CA had a typo in its key size with an 8196-bit RSA key.
BoringSSL is removing support for anything more than 8192-bit RSA
keys in an upcoming change which caused this test and all that rely
on the root CA to fail. This commit updates the root CA with the
intended 8192-bit RSA key and recreates all the intermediate CA and
leaf certificates to support the new certificate chain from this
new root. Where necessary, signatures were regenerated as well with
the new keys. All attributes of the original certificates were kept
as close to the previous value as possible, including serial numbers
that appeared to be explicitly assigned.

Bug: 434905469
Flag: TEST_ONLY
Test: atest CertUtilsTest CertXmlTest SigXmlTest
Change-Id: I126b67fc05c3d8ab1c0142e5d87673940ead4c64