Default element provider for clock regions
This creates an element provider for the clock regions. These are
overriden by any region the clock specifies when the map is constructed,
but keeps the code streamlined in the common case.

This change also improves smartspace rendering by removing the need for
a wrapper FrameLayout in those cases, and moves some padding that was
applied in multiple places. Finally, it also re-enables the layout
callbacks that were previously removed and are used to track the
location of certain on-screen elements.

Bug: 432451019
Bug: 437341450
Bug: 437359698
Test: Checked rendering of clock regions in flexiglass
Flag: com.android.systemui.scene_container
Change-Id: Icdb53149e309899cf6dab90c36704caa498cbdd0