Add flag to sandbox external display rotation for responsive camera apps.
To show upright preview, camera apps need to rotate the image received from the camera feed by the amount the display is rotated - this is a system approximation for how much the camera is rotated, given that camera framework is unaware of device rotation. However, most apps use the display the activity is currently running on, meaning this could be the external display unrelated to the camera rotation. This leads to stretched and sideways previews.

This treatment sandboxes display rotation to use the rotation of the internal display. No camera feed rotate-and-crop and letterboxing should occur, to keep the app as responsive as possible and preserve FOV.

Flag: EXEMPT flag creation
Bug: 395063101
Test: N/A
Change-Id: I754397fcade5859b33a2ba09e0d203d052ba1fef