Flag off LS -> PRIMARY_BOUNCER notificationAlpha.
This transition view model is obsolete when Flexiglass is enabled, as LS -> PRIMARY_BOUNCER is a scene transition.

We're going to need to do this with quite a few transition view models, but I'm going to start with the ones that address specific Fixiglass bugs for now.

Bug: 435151650
Test: swipe up to bouncer
Flag: com.android.systemui.scene_container
Change-Id: Ia604182481a6169bf08f010c6a396ddc3b3c0783