Close PiP dismiss target upon screen off
When screen turns off and there is an ongoing PiP move, the
ACTION_CANCEL motion event is used to close the PiP dismiss target. This
ensures a proper cleanup and there is no dismiss target when the device
is later turned on.

Flag: EXEMPT bugfix
Bug: 408819806
Video: http://recall/-/aaaaaabFQoRHlzixHdtY/biEQbw9rcKkgntsULNmi52
Test: Verify PiP dismiss target is gone when turn on device
Change-Id: If29639b7ff8fca391e9c08b659b7e075fe7f5459