Modify device activity monitor to detect the start of "flight mode".
Currently it only detects when it ends. With this change, the device activity monitor notifies clients when it starts. This can help to know when the fallback needs to stop.

No change in the behavior is done yet. This only provides the capability to detect the beginning of "flight mode" and will later be used in the Fused Time Zone Detector.

go/android-tz-detector
go/ftzd-algo
go/ftzd-cases
go/ftzd-scenarios

Test: atest FrameworksTimeCoreTests
Test: atest FrameworksTimeServicesTests
Test: atest FrameworksTelephonyTests
Flag: android.timezone.flags.enable_fused_time_zone_detector
Bug: 394770805
Change-Id: I3b42dbdc78e40f08c831b454bd3a0dd868f09baa