[Flexiglass] Remove NotificationScrollingStack from QuickSettingsScene
The NotificationScrollingStack is only part of QuickSettingsScene due to legacy,
not due to UX spec. The "drive by" showing of the notifications when closing QS
and going directly to Gone or Lockscreen is an accident of implementation and
not intended spec.

This CL removes the "drive by" animation, and replaces it with simple
STL transitions.

Fixes: 433910331
Bug: 436646848
Test: test all the scenarios for navigate to/from QuickSettingsScene
Flag: com.android.systemui.scene_container
Change-Id: Ib13952f7894f8462872c183b362ddad5a05ac0da