Fix media in landscape in ShadeScene
Fixes the measure policy to properly put QS and media side by side
(before we had a hack because qs lived inside an AndroidView).

This also refactors AnimateQsTilesViewModel and how it's used in
SceneContainer to make it an activatable.

Also, set `usingMediaInCompose` as true in kosmos (but is a var). Most
tests don't actually care or in most cases, media is enabled

Test: manual, open media in landscape in single shade
Test: manual, go from ShadeScene without media to QuickSettingsScene
with media in landscape
Test: atest AnimateQsTilesViewModelTest
Fixes: 436825648
Flag: com.android.systemui.scene_container

Change-Id: I049290d3fb8d85bcf6606267c1b8c6a5a2f6193f