Add Keyboard Backlight Toggle functionality
This commit adds functionality to the toggle keyboard backlight key.
When the key is pressed, if the backlight is on, it will be turned off.
If the backlight is off, it will be restored to the previous level.
With the invocation of the keyboard backlight toggle, ambient light
sensor adjustment will be disabled for the remainder of the session.

Bug: 401626732, 367748270
Test: Build and flash DUT, enable flag, verify that when pressing the
keyboard backlight toggle key, that the keyboard backlight toggles as
expected.
Test: `atest KeyboardBacklightControllerTests`
Flag: com.android.hardware.input.keyboard_backlight_shortcuts

Change-Id: I7bb6a0be80051ed8a50e5525907191c1527f5228