watchdog: Dump virtualizationservice on ANR
When system_server makes a binder call into virtualizationservice, it
can hang, causing a system-wide ANR. The resulting backtrace in the
bugreport only shows that system_server is stuck in a binder
transaction, but provides no information on if or why
virtualizationservice is unresponsive.

To enable better debugging of these hangs, this change adds the
virtualizationservice process to the Watchdog's list of critical
processes to monitor. This ensures that when an ANR occurs, Watchdog
will dump the backtrace of virtualizationservice.

The corresponding SELinux permission is also added to allow
system_server to perform the dump on the virtualizationservice
process.

Flag: EXEMPT add AVF's core process to Watchdog
Bug: 430929909
Bug: 437843268
Test: m
Change-Id: I7294e28e0549144735bc3195e9dfcd9208c84c20