Fix restoring IME visibility on profile switch
Currently when switching from one user (profile) to another, we are
hiding the IME (by setting the requestedVisibleTypes) on the WindowState
of the user we switch to (instead of the one we switch from). This makes
it so the IME is always hidden when going from an app on the personal
profile to one on the work profile (and vice versa).

This makes sure we send a hide signal to the IME of the user we switch
away from, without changing the requestedVisibleTypes of any IME client.

Moreover, this fixes an issue when having apps from two different
profiles in split screen, both with the IME requested visible. If the
IME is hidden on app1, user switches to app2, then back to app1, the IME
can't be requested anymore due to a race condition between reading the
requestedVisibleTypes of the RemoteInsetsControlTarget (split screen),
and updating the ImeInputTarget from app2 to app1. To fix this, the
ImeInputTarget is reset when the IME window is removed, such that the
RemoteInsetsControlTarget requestedVisibleTypes are also reset.

Flag: EXEMPT BUGFIX
Test: atest MultiUserMockImeTest#testProfileSwitchingCanRestoreImeVisibility
Test: manual, open one personal profile app and one work profile app in
  split screen, show IME on both, hide IME on the first, switch to the
  second, switch back to the first and tap on an EditText to show.
Bug: 394783004
Change-Id: I813d5a587b46f13ef98c0cf88f4553aa0a86ee95