[Flexiglass] Fix DeviceEntryIcon background's alpha being shown as white circle on AOD
Issue - When transitioning from some scenes in Keyguard, the
DeviceEntryIcon's background white circle's alpha is computed as 1f and
as a result a white circle is shown on AOD as a background for
DeviceEntryIcon.

Fix - Set the DeviceEntryIcon's background icon's alpha is computed
based on Keyguard state and current scene.

Bug: 435672922
Test: Manual, tried transitioning between flexiglass
scenes in which the background circle was being shown as white (i.e.,
from Lockscreen to AOD or Lockscreen to Dream), this change fixes that
issue.
Flag: com.android.systemui.scene_container

Change-Id: Ie0d3c4e705cc58f3b579d9a5f2c07ea37594a09b