Expose the rendering surface control for use by DisplayOffload
If attempting to offload a WatchFace (e.g. wallpaper) then
DisplayOffload must track the layerID of the surface control.
This is to ensure that it only permits the offload to occur
while the layer associated with that ID is visible on screen.

Test: atest WearFrameworksHalV3IntegrationTests
Flag: NONE The usage of this getter is flagged with com.android.clockwork.flags.render_session_service
Bug: 436487952
Change-Id: I18feafd06bda2b6a015c9ab2c1333c6f910054e2